Moved entity Delta utilities into utils/delta.ts
